Changed the folder structure of the plugin to make it more readable
