+++ src/main/java/io/jenkins/plugins/VirtualPipeline/VirtualPipelineRecorder.java
@@ -19,6 +19,14 @@ import hudson.model.*;
 import hudson.tasks.BuildStepDescriptor;
 import hudson.tasks.Publisher;
 import hudson.tasks.Recorder;
+import io.jenkins.plugins.VirtualPipeline.actions.VirtualPipelineHTMLAction;
+import io.jenkins.plugins.VirtualPipeline.actions.VirtualPipelineHistoryDiffAction;
+import io.jenkins.plugins.VirtualPipeline.actions.VirtualPipelineOffsetAction;
+import io.jenkins.plugins.VirtualPipeline.actions.VirtualPipelineProjectAction;
+import io.jenkins.plugins.VirtualPipeline.input.VirtualPipelineInput;
+import io.jenkins.plugins.VirtualPipeline.input.VirtualPipelineInputDescriptor;
+import io.jenkins.plugins.VirtualPipeline.model.LineWithOffset;
+import io.jenkins.plugins.VirtualPipeline.model.VirtualPipelineLineOutput;
 import jenkins.model.Jenkins;
 import jenkins.tasks.SimpleBuildStep;
 import org.jenkinsci.Symbol;
