+++ src/main/java/io/jenkins/plugins/VirtualPipeline/actions/VirtualPipelineHTMLAction.java
@@ -10,13 +10,15 @@
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package io.jenkins.plugins.VirtualPipeline;
+package io.jenkins.plugins.VirtualPipeline.actions;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import hudson.console.ConsoleNote;
 import hudson.model.Action;
 import hudson.model.Run;
+import io.jenkins.plugins.VirtualPipeline.model.LineType;
+import io.jenkins.plugins.VirtualPipeline.model.VirtualPipelineLineOutput;
 import jenkins.tasks.SimpleBuildStep;
 
 import java.io.BufferedReader;
