+++ src/main/java/io/jenkins/plugins/VirtualPipeline/input/VirtualPipelineInput.java
@@ -10,13 +10,29 @@
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package io.jenkins.plugins.VirtualPipeline;
+package io.jenkins.plugins.VirtualPipeline.input;
 
-public enum LineType {
-    ONE_LINE, // for lines generated by single input
-    START_MARK, // for lines matching startMark
-    CONTENT_LINE, // for lines that are part of the advanced regex content
-    END_MARK, // for lines matching startMark
-    LIMIT_REACHED_LINE, // same as CONTENT_LINE, but marking reached limit of lines
-    DEFAULT // used for lines with no marking meaning
+import hudson.Extension;
+import hudson.ExtensionPoint;
+import hudson.model.Describable;
+import hudson.model.Descriptor;
+import jenkins.model.Jenkins;
+
+public abstract class VirtualPipelineInput implements Describable<VirtualPipelineInput>, ExtensionPoint {
+    //TODO change to protected
+    public VirtualPipelineInput() {
+
+    }
+
+    @Override
+    public Descriptor<VirtualPipelineInput> getDescriptor() {
+        return Jenkins.get().getDescriptorOrDie(this.getClass());
+    }
+
+    @Extension
+    public static class DescriptorImpl extends Descriptor<VirtualPipelineInput> {
+        public String getDisplayName() {
+            return "";
+        }
+    }
 }
