+++ src/main/java/io/jenkins/plugins/VirtualPipeline/model/LineType.java
@@ -10,9 +10,13 @@
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package io.jenkins.plugins.VirtualPipeline;
+package io.jenkins.plugins.VirtualPipeline.model;
 
-import hudson.model.Descriptor;
-
-public class VirtualPipelineInputDescriptor extends Descriptor<VirtualPipelineInput> {
+public enum LineType {
+    ONE_LINE, // for lines generated by single input
+    START_MARK, // for lines matching startMark
+    CONTENT_LINE, // for lines that are part of the advanced regex content
+    END_MARK, // for lines matching startMark
+    LIMIT_REACHED_LINE, // same as CONTENT_LINE, but marking reached limit of lines
+    DEFAULT // used for lines with no marking meaning
 }
