+++ src/main/resources/io/jenkins/plugins/VirtualPipeline/actions/VirtualPipelineHTMLAction/index.jelly
