+++ src/main/resources/io/jenkins/plugins/VirtualPipeline/actions/VirtualPipelineProjectAction/summary.jelly
