+++ src/main/resources/io/jenkins/plugins/VirtualPipeline/input/VirtualPipelineInputAdvanced/config.jelly
