+++ src/main/resources/io/jenkins/plugins/VirtualPipeline/input/VirtualPipelineInputSimple/config.jelly
